weather_api_1/requirements.txt
bcrypt==3.1.7
beautifulsoup4==4.11.1
certifi==2020.6.20
chardet==4.0.0
click==8.1.3
cryptography==3.3.2
dbus-python==1.2.16
distro==1.5.0
dnspython==2.0.0
Flask==2.1.2
gunicorn==20.1.0
httplib2==0.18.1
idna==2.10
importlib-metadata==4.11.4
itsdangerous==2.1.2
jeepney==0.6.0
Jinja2==3.1.2
MarkupSafe==2.1.1
musicbrainzngs==0.7.1
mutagen==1.45.1
psutil==5.8.0
pycairo==1.16.2
pycups==2.0.1
pycurl==7.43.0.6
PyGObject==3.38.0
pyinotify==0.9.6
pyOpenSSL==20.0.1
PySimpleSOAP==1.16.2
python-debian==0.1.39
python-debianbts==3.1.0
python-gnupg==0.4.6
pythondialog==3.5.1
pytz==2022.1
pyxdg==0.27
quodlibet==4.3.0
reportbug==7.10.3+deb11u1
requests==2.25.1
requests-toolbelt==0.9.1
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.2.post1
systemd-python==234
urllib3==1.26.5
Werkzeug==2.1.2
xdg==5
zipp==3.8.0
